pyproject.toml
[project]
name = "crazyflie-fpv-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cflib==0.1.30",
    "einops>=0.8.2",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "huggingface-hub>=1.4.1",
    "importlib-resources==6.5.2",
    "libusb-package==1.0.26.3",
    "numpy==2.4.2",
    "opencv-contrib-python>=4.13.0.92",
    "packaging==25.0",
    "pipreqs>=0.5.0",
    "pygame==2.6.1",
    "pyusb==1.3.1",
    "scipy==1.17.0",
    "timm>=1.0.24",
    "torchvision>=0.25.0",
    "ultralytics>=8.3.0",
    "wandb>=0.25.0",
]
